+++ paper/paper.bib
@@ -472,6 +472,7 @@
 	url = {https://github.com/jaclark5/mapsci},
 	author = {Clark, Jennifer A and Santiso, Erik E.},
 	year = {2020},
+	doi = {10.5281/zenodo.4271975},
 }
 
 @article{herdes_coarse_2015,
