### 2024-09-18 #7090 @JacksonBurns
@srmnitc yes would be happy to review. I have reviewed in the past ([profile here](https://reviewers.joss.theoj.org/reviewers/3341)) and this looks well within my experience.

Love the name of the package, @jaclark5
