### 2024-10-18 #7365 @editorialbot
[REVIEW]: DESPASITO: A Python Package for SAFT EOS Parametrization and Thermodynamic Calculations
**Submitting author:** @jaclark5 (<a href="http://orcid.org/0000-0003-4897-5651">Jennifer A Clark</a>)
**Repository:** https://github.com/Santiso-Group/despasito
**Branch with paper.md** (empty if default branch): master
**Version:** 0.3.0
