### 2024-10-28 #7365 @jaclark5
> @jaclark5 I am going through the author list and had a quick question. I see you and your co-author Andrew on the committers list on the repo. I assume Erik is your PI. Could you clarify Nathan's contribution to the project, since I don't see his user in the GitHub committers (see: https://github.com/Santiso-Group/despasito/graphs/contributors)?

@JacksonBurns, thanks for the question. Nathan wrote a first, "functional" draft of this program. At this point I rewrote all files and implemented OOP, with the exception being `despasito.equations_of_state.saft.compiled_modules.ext_gamma_mie_python.py` where I did make some changes but it is largely written by Nathan; the numba and cython versions of that file are closely derived.
