### 2024-11-02 Santiso-Group/despasito#33 @yuzie007
[JOSS Review] Inconsistency between documentation and the PyPI source
(This is a part of my review for JOSS https://github.com/openjournals/joss-reviews/issues/7365. cc: @srmnitc)

Dear @jaclark5 and other authors of DESPASITO,

I am Yuji Ikeda @yuzie007 and reviewing DESPASITO for JOSS.
Thank you for writing a wonderful code and making it open-source.

Following the JOSS guideline, I am first checking the consistency with the documentation and the actual code in the PyPI source code. Could you check the points below and, when necessary, update the repo?

I will also (1) test other examples and (2) read the paper submitted to JOSS. The feedback for these will be reported in other issues.

- [x] **Functionality** – **Installation**: In [`README.md`](https://github.com/Santiso-Group/despasito/blob/0851b58d874175cb11afc82bf2622c1dd474f617/README.md?plain=1#L34) and [`getting_started.rst`](https://github.com/Santiso-Group/despasito/blob/0851b58d874175cb11afc82bf2622c1dd474f617/docs/getting_started.rst?plain=1#L10), the available Python version is written up to 3.8, which has already been [its end-of-life](https://devguide.python.org/versions/#python-release-cycle). The authors should update check and update the Python versions that are available for DESPASITO and not yet their end-of-life.

- [x] **Functionality** – **Installation**: In `getting_started.rst`, the authors write ["Option 3: Install locally with python"](https://github.com/Santiso-Group/despasito/blob/0851b58d874175cb11afc82bf2622c1dd474f617/docs/getting_started.rst?plain=1#L23-L27). This looks for me redundant to "Option 2: Install locally with pip".
I am not sure if we need to recommend users Option 3. In the modern Python they use `pyproject.toml` rather than `setup.py`, and probably Option 3 does not work with `pyproject.toml`.

- [x] **Documentation** – **Automated tests**: [In the present GitHub Actions workflow, the automated tests are done only up to Python 3.10](https://github.com/Santiso-Group/despasito/blob/0851b58d874175cb11afc82bf2622c1dd474f617/.github/workflows/CI.yaml#L28). Likely the authors can update this to check up to which latest Python versions are available for DESPASITO.

- [x] **Documentation** – **Example usage**: While in [`settingup.rst`](https://github.com/Santiso-Group/despasito/blob/master/docs/settingup.rst?plain=1#L86) the output file name of the first example is written as `despasito_out.txt`, I actually obtained`out_vapor.txt`. Please cross check and, if reasonable, revise this.

- [x] **Documentation** – **Example usage**: In `settingup.rst`, [the example usage as a python package is given as `hexane_heptane_test.txt`](https://github.com/Santiso-Group/despasito/blob/master/docs/settingup.rst?plain=1#L97-L129). However, the content is inconsistent with `despasito/examples/saft_gamma_mie/hexane_heptane_activity_coeff/hexane_heptane_test.py`. Further, both the script shows error in the directory `despasito/examples/saft_gamma_mie/hexane_heptane_activity_coeff` as
```
Traceback (most recent call last):
  File "/Users/ikeda/miniforge3/envs/myenv/lib/python3.11/site-packages/despasito/thermodynamics/__init__.py", line 61, in thermo
    output_dict = func(Eos, **kwargs)
                  ^^^^^^^^^^^^^^^^^^^
  File "/Users/ikeda/miniforge3/envs/myenv/lib/python3.11/site-packages/despasito/thermodynamics/calculation_types.py", line 1043, in vapor_properties
    raise ValueError("Ambiguous mixture composition. Define yilist")
ValueError: Ambiguous mixture composition. Define yilist

Traceback (most recent call last):
  File "/Users/ikeda/codes/despasito/despasito/examples/saft_gamma_mie/hexane_heptane_activity_coeff/test.py", line 18, in <module>
    output = thermo.thermo(
             ^^^^^^^^^^^^^^
  File "/Users/ikeda/miniforge3/envs/myenv/lib/python3.11/site-packages/despasito/thermodynamics/__init__.py", line 63, in thermo
    raise TypeError("The calculation type, '{}', failed".format(calculation_type))
TypeError: The calculation type, 'vapor_properties', failed
```
and
```
Traceback (most recent call last):
  File "/Users/ikeda/codes/despasito/despasito/examples/saft_gamma_mie/hexane_heptane_activity_coeff/hexane_heptane_test.py", line 8, in <module>
    despasito.initiate_logger(console=True, verbose=10)
  File "/Users/ikeda/miniforge3/envs/myenv/lib/python3.11/site-packages/despasito/__init__.py", line 95, in initiate_logger
    handler_logfile.close()
    ^^^^^^^^^^^^^^^^^^^^^
AttributeError: 'NoneType' object has no attribute 'close'
```

- [x] **Documentation** – **Example usage**: In [`startfitting.rst`](https://github.com/Santiso-Group/despasito/blob/0851b58d874175cb11afc82bf2622c1dd474f617/docs/startfitting.rst?plain=1#L17), it is written that the `input_fit.json` for methanol can be found in the `examples` directory. However, I could only found it fro another system in `examples/saft_gamma_mie/butane_solubility`. (I could run `input_fit.json` in this directory.)
