### 2024-11-03 Santiso-Group/despasito#34 @yuzie007
[JOSS Review] Suggestions on `paper.md` etc.
(This is a part of my review for JOSS https://github.com/openjournals/joss-reviews/issues/7365. cc: @srmnitc)

As a follow-up to #33, I suggest the following for `paper.md`.

- [ ] **Software paper** – **Summary**: PyPi -> PyPI (The last "I" should be a capital letter.)
- [ ] Throughout the manuscript, in the present citation style (not in the superscript style but in the bracket style) citations should come before punctuations, not like `,[@plunkett_plunketts_2017]` but like `[@plunkett_plunketts_2017],`.
- [ ] **Software paper** – **A statement of need**: Citations for the "PC-SAFT" theory / methodology may be appreciated.
- [ ] **Section 2.1**: Citations for [MolSSI Cokkiecutter](https://github.com/MolSSI/cookiecutter-cms?tab=readme-ov-file#acknowledgments) may be appreciated.
- [ ] **Section 2.3**: While in the manuscript the currently available version is written as 0.2.0, [in CHANGELOG.md, this is now 0.3.0](https://github.com/Santiso-Group/despasito/blob/0851b58d874175cb11afc82bf2622c1dd474f617/CHANGELOG.md?plain=1#L8). This inconsistency should be solved.
- [ ] **Section 3 & References**: The authors can provide DOI for MAPSCI ([10.5281/zenodo.4271975](https://doi.org/10.5281/zenodo.4271975) or [10.5281/zenodo.4271975](https://doi.org/10.5281/zenodo.4271975)) together with its publication year.
- [ ] **References**: The hyperlink for the software PC-SAFT (Baird, Z. (2021)) is outdated. The authors can refer to the latest PyPI page (https://pypi.org/project/pcsaft/) or the GitHub page (https://github.com/zmeri/PC-SAFT).
- [ ] **Section 2.3**: As written here, "Second derivative properties are a priority in the next release", probably meaning that this is not yet available. As far as I learned from [the paper of Papaioannou et al. in (2014)](https://doi.org/10.1063/1.4851455), this feature is what users would wish. If this is not yet implemented, this information may be better to be given somewhere in the actual code document in `docs` as well.
