### 2024-11-11 Santiso-Group/despasito#33 @JacksonBurns
Hi I am the other reviewer. I also noticed the inconsistent listing of support Python versions between the README and the CI, but also in the `setup.py` file, which allows any modern version:
```python
    python_requires=">=3.6",
```
