### 2024-11-11 284f0ab
Allow installation without numpy
When installing `despasito`, the user must have `numpy` installed for the import to run. However, if the user does not have Cython installed, `numpy` will not be used anyway.

I propose moving the `numpy` import inside the same optional logic as Cython, and modifying the warning to be more explicit.

If you like this change, the README should also be updated to explain how `numpy` and Cython are optional for acceleration.

This PR is part of an ongoing JOSS review at https://github.com/openjournals/joss-reviews/issues/7365
