### 2024-11-11 Santiso-Group/despasito#36 @JacksonBurns
[BUG]: Peng-Robinson Acetone+Chloroform Example Fails
**Describe the bug**
The Peng-Robinson example fails, producing the following output:
```bash
(despasito_env) jackson@jackson-Precision-7540:~/despasito/despasito/examples/cubic_peng_robinson/acetone_chloroform$ cat out_xi.txt
# This data was generated in DESPASITO using the thermodynamic
# T [K], xi1, xi2, P [Pa], yi1, yi2, flagl, flagv, obj,
 335.75, 0.098, 0.902, nan, nan, nan, 3, 3, nan,
 336.65, 0.186, 0.814, nan, nan, nan, 3, 3, nan,
 337.25, 0.266, 0.734, nan, nan, nan, 3, 3, nan,
 337.55, 0.36, 0.64, nan, nan, nan, 3, 3, nan,
 336.95, 0.468, 0.532, nan, nan, nan, 3, 3, nan,
 335.85, 0.578, 0.422, nan, nan, nan, 3, 3, nan,
 334.65, 0.673, 0.327, nan, nan, nan, 3, 3, nan,
 333.45, 0.755, 0.245, nan, nan, nan, 3, 3, nan,
 332.15, 0.827, 0.173, nan, nan, nan, 3, 3, nan,
```
with these log messages at the end of execution:
```bash
INFO:despasito.thermodynamics.calc:Obj Func: -0.00027928178452807373, Pset: 100176.90109275474, Pcalc: 100176.90109270591
WARNING:despasito.thermodynamics.calculation_types:T (K), xi: 332.15 [0.827 0.173], calculation did not produce a valid result.
INFO:despasito.thermodynamics.calculation_types:P (Pa), yi: nan [nan nan]
INFO:despasito.thermodynamics.calculation_types:--- Calculation bubble_pressure Complete ---
```

**To Reproduce**
 1. installed despasito in python 3.8 from pip following installation instructions
 2. cloned repository and navigated to example for peng-robinson acetone/chloroform
 3. ran the liquid, vapor, fit_liquid, and xi input files (I had to rename the output of fit_liquid to match that which was expected in xi_input)
 4. Got above result

**Expected behavior**
This is presumably supposed to produce some actual profiles

 - OS: Ubuntu
 - Version 22

**Additional context**
Despasito 0.3.0

This issue is part of an ongoing JOSS review at https://github.com/openjournals/joss-reviews/issues/7365
