### 2024-11-11 Santiso-Group/despasito#37 @JacksonBurns
[General] Add contributing guidelines
**Describe the issue**
The README welcomes contributions, but some more explicit directions can be helpful to new users.

**Expected behavior**
Consider the official GitHub guidelines for writing these guidelines: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors

**Additional context**
This issue is part of an ongoing JOSS review at https://github.com/openjournals/joss-reviews/issues/7365
