### 2024-11-11 Santiso-Group/despasito#37 @JacksonBurns
I see actually that you have some guidelines here: https://github.com/Santiso-Group/despasito/tree/master/devtools#how-to-contribute-changes

Maybe just link to this document in the README, and expand it?
