### 2024-11-11 #7365 @JacksonBurns
@srmnitc @jaclark5 I have completed my initial review. Overall this package is in great shape. Excellent work on the literature review section of your paper, especially.

My most significant outstanding points are linked above this comment: (1) simplified installation (2) bug in a demo and (3) improved contributor guidelines. With these fixed, despasito will be a perfect addition to JOSS.
