### 2024-11-13 a148e48
Address referencing issues. Resolves #34
