### 2024-11-13 Santiso-Group/despasito#34 @jaclark5
@yuzie007 thank you for the feedback. I addressed your concerns in PR #38. The lack of date and doi for MAPSCI seems to have been an error with bibtex formatting. I need to let the JOSS bot compile the paper to see if it is fixed.
