### 2024-11-14 Santiso-Group/despasito#33 @jaclark5
@yuzie007 Thank you for your comments and pointing out the inconsistencies. I've addressed them in #39 if you'd like to view them. Notice that the docs fail due to the update in dependencies required that have been included in PR 38.
