### 2024-11-14 Santiso-Group/despasito#33 @jaclark5
@JacksonBurns I've addressed the inconsistencies between the paper, README, and CI testing. I will keep the setup.py file as is so that installation won't automatically fail with future releases of python, quite possibly without cause.
