### 2024-11-14 366403c
Remove testing python=3.13, not yet supported by conda-incubator in github actions
