### 2024-11-14 Santiso-Group/despasito#34 @jaclark5
@yuzie007 The paper and docs have compiled and it seems like your comments have been addressed and rendered correctly.:
https://github.com/Santiso-Group/despasito/actions/runs/11829116940/artifacts/2185390672
Note the additional sentence in the third paragraph of the documentation:
https://despasito--38.org.readthedocs.build/en/38/
