### 2024-11-14 Santiso-Group/despasito#33 @jaclark5
@yuzie007 all tests have passed for python 3.10, 3.11, and 3.12. python 3.13 is not yet supported by the condo-incubator used in GitHub actions.
