### 2024-11-16 Santiso-Group/despasito#33 @yuzie007
Thank you so much for #39 and information. Regarding Python 3.13, if it cannot be tested just due to the GitHub action issue, I would feel it is OK to also include Python 3.13 also in the support, if it can pass all tests, e.g., in a local environment.

After #38 and #39 are merged, I would like to try remaining tests, and then I think I will finish my JOSS review.
