### 2024-11-17 54cd4d0
Update installation dependencies
