### 2024-12-08 Santiso-Group/despasito#41 @yuzie007
[JOSS Review] Suggestions on `examples` etc.
(This is a part of my review for JOSS https://github.com/openjournals/joss-reviews/issues/7365. cc: @srmnitc)

As the last part of my JOSS review, I tested input files in `examples` and found the following, which can probably fixed.

- [x] `cubic_peng_robinson/acetone_hexane/input_xi_318.json`: `"PRcross_315.json"` and `"out_xi_315.txt"` should probably be modified as `"PRcross_318.json"` and `"out_xi_318.txt"`, respectively.
- [x] `library`: The usage is not clear.
- [x] `saft_gamma_mie/butane_solubility/input_test.json`: The calculation fails with the error message `ImportError: The experimental data type, 'test', was not found`.
- [x] `saft_gamma_mie/decane_helmholtz/run_test.py`: The calculation fails the error message `AttributeError: module 'despasito' has no attribute 'equations_of_state'`.
- [ ] `saft_gamma_mie/hexane_heptane_activity_coeff/test.py`: The calculation fails with the error message `TypeError: The calculation type, 'vapor_properties', failed`.
- [x] `saft_gamma_mie/butane_solubility/input_fit_grid.json`: The calculation fails with the error message `TypeError: The parameter fitting failed`.
- [x] `saft_gamma_mie/Fit_CH3/grid_minimization/input.json`: The calculation fails with the error message `TypeError: The parameter fitting failed`.
- [x] `saft_gamma_mie/Fit_CH3/differential_evolution/input.json`: The calculation seems to take too long time for me to finish even nearly 1 hour. Can you address this in some way? Or, can you at least provide users an expected time?
