### 2024-12-16 Santiso-Group/despasito#41 @jaclark5
I do not see a file: `saft_gamma_mie/hexane_heptane_activity_coeff/test.py`

`saft_gamma_mie/Fit_CH3/differential_evolution/input.json` should now run in ~21 min.

I added to the readme that library contains the parameter files used in the examples.
