### 2024-12-16 Santiso-Group/despasito#41 @yuzie007
> I do not see a file: saft_gamma_mie/hexane_heptane_activity_coeff/test.py

Sorry for confusion; that was what I made based on despasito documentation. You can leave it.
