### 2024-12-16 #7365 @jaclark5
@srmnitc I've addressed the comments of @yuzie007, however, with the deprecation of setup.py files with pip 25 I think I need to make that conversion before moving forward.

I think the changes here warrant another minor version release. Should I make that release and then update the documentation and joss paper to reflect the version change? How is that normally handled?
