### 2024-12-17 #7365 @srmnitc
@jaclark5 feel free to update. We have to make a new release just before acceptance. The version associated with that release will be the one that appears in the JOSS paper.
