### 2024-12-18 cc3fbe0
Convert setup.py to toml
