+++ src/despasito/_version.py
@@ -1,0 +1,1 @@
+__version__ = "0.3.0+15.g6e200c7.dirty"
