+++ setup.py
@@ -4,39 +4,35 @@ DESPASITO: Determining Equilibrium State and Parametrization Application for SAF
 """
 
 import os
-from setuptools import Extension, setup
 import glob
-import numpy as np
+from setuptools import Extension, setup
 
-short_description = __doc__.split("\n")
+
+import numpy as np
+from Cython.Build import cythonize
+
 fpath = os.path.join("despasito", "equations_of_state", "saft", "compiled_modules")
 extensions = []
 
-try:
-    from Cython.Build import cythonize
-    flag_cython = True
-except Exception:
-    print(
-        'Cython not available on your system. Dependencies will be run with numba.'
+cython_list = glob.glob(os.path.join(fpath, "*.pyx"))
+for cyext in cython_list:
+    name = os.path.split(cyext)[-1].split(".")[-2]
+    cy_ext_1 = Extension(
+        name=os.path.join(fpath,name).replace(os.sep,"."), 
+        sources=[cyext], 
+        include_dirs=[fpath, np.get_include()],
+        )
+    extensions.extend(
+        cythonize(
+            [cy_ext_1],
+            compiler_directives={
+                'language_level': 3,
+                'cdivision': False,
+                "boundscheck": True
+            }))
+
+if __name__ == "__main__":
+    setup(
+        name='despasito',
+        ext_modules=extensions,
     )
-    flag_cython = False
-
-if flag_cython:
-    cython_list = glob.glob(os.path.join(fpath, "*.pyx"))
-    for cyext in cython_list:
-        name = os.path.split(cyext)[-1].split(".")[-2]
-        cy_ext_1 = Extension(
-            name=name, 
-            sources=[cyext], 
-            include_dirs=[fpath, np.get_include()],
-            )
-        extensions.extend(
-            cythonize(
-                [cy_ext_1],
-                compiler_directives={
-                    'language_level': 3,
-                    'cdivision': False,
-                    "boundscheck": True
-                }))
-
-setup(ext_modules=extensions)
