+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools>=61.0", "versioningit~=2.0", "cython", "numpy"]
+requires = ["setuptools>=61.0", "numpy >= 1.15", "versioningit~=2.0", "wheel", "Cython>=0.29"]
 build-backend = "setuptools.build_meta"
 
 [project]
@@ -71,7 +71,7 @@ match = ["*"]
 default-tag = "1.0.0"
 
 [tool.versioningit.write]
-file = "src/despasito/_version.py"
+file = "despasito/_version.py"
 
 [tool.black]
 line-length = 120
