+++ despasito/parameter_fitting/__init__.py
@@ -138,7 +138,8 @@ def fit(
             ]
         else:
             raise ValueError(
-                f"'fit_parameter_names' must be a list not: {optimization_parameters["fit_parameter_names"]}")
+                f"'fit_parameter_names' must be a list not: {optimization_parameters['fit_parameter_names']}"
+            )
         
     # Generate initial guess and bounds for parameters if none was given
     optimization_parameters = ff.consolidate_bounds(optimization_parameters).copy()
