### 2025-01-01 dc17ad2
Remove versioneer
