+++ despasito/_version.py
@@ -1,1 +1,0 @@
-__version__ = "0.3.0+28.ge5ffa36.dirty"
