+++ MANIFEST.in
@@ -1,6 +1,5 @@
 include LICENSE
 include MANIFEST.in
-include versioneer.py
 
 graft despasito
 #global-exclude *.py[cod] __pycache__ *.so
