### 2025-01-01 #7365 @jaclark5
@srmnitc I just pushed the new release to zenodo. I think we are good to go.
