### 2025-01-09 #7365 @srmnitc
@yuzie007 Thanks a lot for your review! Could you please confirm that you recommend publication of this package in JOSS?
