### 2025-01-09 #7365 @yuzie007
@srmnitc Yes, I recommend publication of DESPASITO in JOSS!
