### 2025-01-30 #7365 @srmnitc
@jaclark5 could you please take a look at the above comment, and finish these tasks:

Double check authors and affiliations (including ORCIDs)
Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
Archive the release on Zenodo/figshare/etc and post the DOI here.
Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
Make sure that the license listed for the archive is the same as the software license.
If you have any questions, please feel free to ask.
