+++ paper/paper.bib
@@ -463,7 +463,7 @@
 	url = {https://github.com/Santiso-Group/despasito},
 	author = {Clark, Jennifer A and Duff, Nathan and Abi-Mansour, Andrew and Santiso, Erik E.},
 	year = {2024},
-	doi = {10.5281/zenodo.13119190},
+	doi = {10.5281/zenodo.14585722},
 }
 
 @misc{clark_mapsci_2020,
