+++ paper/paper.md
@@ -12,8 +12,10 @@ authors:
     affiliation: 1
     corresponding: true
   - name: Nathan Duff
+    orcid: 0000-0003-4070-8989 
     affiliation: 1
   - name: Andrew Abi-Mansour
+    orchid: 0009-0003-9026-5415
     affiliation: 2
   - given-names: Erik E. Santiso
     orcid: 0000-0003-1768-8414
@@ -24,21 +26,21 @@ affiliations:
    index: 1
  - name: The Molecular Sciences Software Institute, Blacksburg, VA, 24060, USA
    index: 2
-date: 05 May 2024
+date: 01 February 2025
 bibliography: paper.bib
 ---
 
 # Summary
 
-DESAPSITO is an open-source Python package that provides access to complex thermodynamic calculations, with a special interest in the statistical associating fluid theory (SAFT) equation of state. The modular design of DESPASITO provides an extensible architecture that could be easily customized or extended by other researchers. The currently supported equations of state (SAFT-$\gamma$-Mie, SAFT-$\gamma$-SW, and Peng-Robinson) thoroughly represent the dynamic use of this platform and its intentional design for expansion. This growth mindset extends into the other modules, as additional thermodynamic calculations and parameter fitting routines may be included. Nonetheless, the current state of DESPASITO can be employed by the simulation community for coarse-grained forcefield development with SAFT-$\gamma$-Mie. The DESPASITO package is available from PyPI and its source code is hosted on GitHub, which provides a platform for community-driven contributions and feedback.
+DESAPSITO is an open-source Python package that provides access to complex thermodynamic calculations, with a special interest in the statistical associating fluid theory (SAFT) equation of state. The modular design of DESPASITO provides an extensible architecture that could be easily customized or extended by other researchers. The currently supported equations of state (SAFT-$\gamma$-Mie, SAFT-$\gamma$-SW, and Peng-Robinson) thoroughly represent the dynamic use of this platform and its intentional design for expansion. This growth mindset extends into the other modules, as additional thermodynamic calculations and parameter fitting routines may be included. Nonetheless, the current state of DESPASITO can be employed by the simulation community for coarse-grained force field development with SAFT-$\gamma$-Mie. The DESPASITO package is available from PyPI and its source code is hosted on GitHub, which provides a platform for community-driven contributions and feedback.
 
 # 1. Statement of Need
 
 To improve on the 20+ years needed between material development and market use [@plunkett_plunketts_2017], theoretical studies have accelerated material design endeavors [@lee_review_2019; @fermeglia_multiscale_2020]. Of the theoretical methods available, coarse-grained molecular dynamics (CG-MD) simulations have gained attention for their ability to access transport and mechanical properties, as well as phase separations of materials [@nikkhah_molecular_2021; @miwatani_performance_2020; @lee_multiscale_2009; @joshi_review_2020]. CG methods represent an atomistic ensemble as a single “bead” that must sufficiently capture the degrees of freedom of the underlying structure to produce realistic results [@fermeglia_multiscale_2020]. Although CG-MD yields computational efficiency in comparison to all-atom molecular dynamics (AA-MD), generating parameters that are representative of the system’s physics is a continuing struggle. An appealing method of obtaining parameters entails a top-down approach to reproduce some macroscopic property in a simulation. Comparison to macroscopic properties entails an iterative procedure which is time consuming, as many lengthy simulations are needed. 
 
[…]
