### 2025-02-02 #7365 @jaclark5
@srmnitc  I've made the final version release, version 0.4.1,
I've also updated the Zenodo archive and made sure that the author list and title match [DOI: 10.5281/zenodo.14788353](https://doi.org/10.5281/zenodo.14788353).
Let me know if there's anything else I can do!
