### 2025-02-02 ace4600
Update CI to generate paper and fix orcid
