+++ .github/workflows/CI.yaml
@@ -98,8 +98,8 @@ jobs:
         with:
           name: paper
           path: paper/paper.pdf
-#      - name: Commit PDF to repository
-#        uses: EndBug/add-and-commit@v9
-#        with:
-#          message: '(auto) Paper PDF Draft'
-#          add: 'paper/*.pdf'
+      - name: Commit PDF to repository
+        uses: EndBug/add-and-commit@v9
+        with:
+          message: '(auto) Paper PDF Draft'
+          add: 'paper/*.pdf'
