### 2025-02-02 Santiso-Group/despasito#43 @srmnitc
[General] Changes for JOSS paper
This is part of JOSS review over at https://github.com/openjournals/joss-reviews/issues/7365

Important points:

- There are references to tables, where are the tables?
- ~~gamma symbols are missing in a number of places, I have listed the ones I found below, but please check again~~
- Please ensure that journal names are either in full, or abbreviated to be consistent

Few recommendations for the paper:

- ~~Please check ORCids of Nathan Duff and Abi-Mansour~~
- ~~line 10-13: Would the mentioned equations need citations? (ah I see it comes later..)~~
- line 33: might need a citation
- ~~line 34: gamma sign between SAFT-gamma-Mie missing~~
- line 56: extra comma
- line 124: maybe a hyperlink to docs would benefit the reader
- line 131: please make sure this confirms to the archive version
- ~~line 144: missign gamma symbols~~
- line 161: a hyperlink to the scipy function could be added
- line 169: comma not needed?
- line 170: maybe links
- line 186: With these three core functions, the major features - comma
- ~~line 222: missing gamma symbol~~
- ~~line 246 - gamma symbol~~
- line 259 - maybe add the reference
- line 288, 338, 344, 347, 353, 384, 419,  - number/issue missing
- line 338 - number/issue missing
