+++ paper/paper.bib
@@ -3,9 +3,20 @@
 	title = {Plunkett's Nanotechnology \& MEMS Industry Almanac 2017},
 	isbn = {978-1-62831-439-7},
 	language = {English},
-	publisher = {PLUNKETT RESEARCH},
+	publisher = {Plunkett Research},
 	author = {Plunkett, Jack W},
 	year = {2017},
+}
+
+@article{thaler_learning_2021,
+	title = {Learning Neural Network Potentials from Experimental Data via Differentiable Trajectory Reweighting},
+	volume = {12},
+	doi = {10.1038/s41467-021-27241-4},
+	pages = {6884},
+	number = {1},
+	year = {2021},
+	journaltitle = {Nature Communications},
+	author = {Thaler, Stephan and Zavadlav, Julija},
 }
 
 @article{lee_review_2019,
@@ -58,6 +69,7 @@
 @article{lee_multiscale_2009,
 	title = {Multiscale Modeling of Dendrimers and Their Interactions with Bilayers and Polyelectrolytes},
 	volume = {14},
+    number = {1},
 	doi = {10.3390/molecules14010423},
 	language = {en},
 	journal = {Molecules},
@@ -432,7 +444,7 @@
 	doi = {10.1021/ie4033999},
 	language = {en},
 	number = {6},
-	journal = {Ind. Eng. Chem. Res.},
+	journal = {Industrial \& Engineering Chemistry Research},
 	author = {Bell, Ian H. and Wronski, Jorrit and Quoilin, Sylvain and Lemort, Vincent},
 	year = {2014},
 	pages = {2498},
@@ -459,11 +471,11 @@
 }
 
 @misc{clark_despasito_2021,
-	title = {DESPASITO: Determining Equilibrium State and Parametrization Application for SAFT, Intended for Thermodynamic Output},
+	title = {DESPASITO: A Python Package for SAFT EOS Parametrization and Thermodynamic Calculations},
 	url = {https://github.com/Santiso-Group/despasito},
 	author = {Clark, Jennifer A and Duff, Nathan and Abi-Mansour, Andrew and Santiso, Erik E.},
-	year = {2024},
-	doi = {10.5281/zenodo.14585722},
+	year = {2025},
+	doi = {10.5281/zenodo.14788353},
 }
 
 @misc{clark_mapsci_2020,
@@ -488,6 +500,7 @@
 @article{herdes_predicting_2016,
 	title = {Predicting the Adsorption of n-Perfluorohexane in BAM-P109 Standard Activated Carbon by Molecular Simulation using SAFT-$\gamma$ Mie Coarse-Grained Force Fields},
 	volume = {34},
+	number = {1},
 	doi = {10.1177/0263617415619528},
 	language = {en},
 	journal = {Adsorption Science \& Technology},
