+++ paper/paper.md
@@ -36,11 +36,11 @@ DESAPSITO is an open-source Python package that provides access to complex therm
 
 # 1. Statement of Need
 
-To improve on the 20+ years needed between material development and market use [@plunkett_plunketts_2017], theoretical studies have accelerated material design endeavors [@lee_review_2019; @fermeglia_multiscale_2020]. Of the theoretical methods available, coarse-grained molecular dynamics (CG-MD) simulations have gained attention for their ability to access transport and mechanical properties, as well as phase separations of materials [@nikkhah_molecular_2021; @miwatani_performance_2020; @lee_multiscale_2009; @joshi_review_2020]. CG methods represent an atomistic ensemble as a single “bead” that must sufficiently capture the degrees of freedom of the underlying structure to produce realistic results [@fermeglia_multiscale_2020]. Although CG-MD yields computational efficiency in comparison to all-atom molecular dynamics (AA-MD), generating parameters that are representative of the system’s physics is a continuing struggle. An appealing method of obtaining parameters entails a top-down approach to reproduce some macroscopic property in a simulation. Comparison to macroscopic properties entails an iterative procedure which is time consuming, as many lengthy simulations are needed. 
+To improve on the 20+ years needed between material development and market use [@plunkett_plunketts_2017], theoretical studies have accelerated material design endeavors [@lee_review_2019; @fermeglia_multiscale_2020]. Of the theoretical methods available, coarse-grained molecular dynamics (CG-MD) simulations have gained attention for their ability to access transport and mechanical properties, as well as phase separations of materials [@nikkhah_molecular_2021; @miwatani_performance_2020; @lee_multiscale_2009; @joshi_review_2020]. CG methods represent an atomistic ensemble as a single “bead” that must sufficiently capture the degrees of freedom of the underlying structure to produce realistic results [@fermeglia_multiscale_2020]. Although CG-MD yields computational efficiency in comparison to all-atom molecular dynamics (AA-MD), generating parameters that are representative of the system’s physics is a continuing struggle. An appealing method of obtaining parameters entails a top-down approach to reproduce some macroscopic property in a simulation. Comparison to macroscopic properties entails an iterative procedure which is time consuming, as many lengthy simulations are needed [@thaler_learning_2021]. 
 
[…]
