### 2025-02-02 Santiso-Group/despasito#43 @jaclark5
Reference to table removed.
New reference provided.
Appropriate hyperlinks added
Appropriate addition and removal of commas was done
Abbreviated journal title for "Industrial & Engineering Chemistry Research" was resolved.
"line 259 - maybe add the reference", these verification tests were not published
Citations for line 288, 338, 347, 384, 419, do not have issue numbers
Citation for line 338 does not have an issue number
