+++ paper/paper.bib
@@ -15,6 +15,7 @@
 	pages = {6884},
 	number = {1},
 	year = {2021},
+	language = {en},
 	journaltitle = {Nature Communications},
 	author = {Thaler, Stephan and Zavadlav, Julija},
 }
