### 2025-02-04 #7365 @srmnitc
> ⚠️ [Error preparing paper acceptance](https://github.com/openjournals/joss-papers/actions/runs/13131436991?check_suite_focus=true). The generated XML metadata file is invalid.
>
> Element surname: [facet 'minLength'] The value has a length of '0'; this underruns the allowed minimum length of '1'.
> Element surname: [facet 'pattern'] The value '' is not accepted by the pattern '[^\d\?]*[^\?\s]+[^\d]*'.

@kyleniemeyer do you have any idea what could be the issue here? I would appreciate the help!
