### 2025-02-04 #7365 @srmnitc
@jaclark5 I am not sure what the issue is, but maybe 'given-names' instead of 'name' in the metadata is the issue, could you change this? We could try again then
