### 2025-02-04 #7365 @kyleniemeyer
@jaclark5 please remove `paper.pdf` from the repository, as this can conflict with our build system. And, I just noticed the last author has the field `names` rather than `name`.
