### 2025-02-05 f3c71f9
Update paper references, pub prep
