### 2025-02-06 #7365 @srmnitc
@jaclark5 `dominik_modeling_2005` seems to appear twice in the bib file, deleting one of them should be final hurdle!
