### 2025-02-06 b547819
Remove redundant reference
