+++ paper/paper.bib
@@ -579,18 +579,6 @@
 	pages = {1244},
 }
 
-@article{dominik_modeling_2005,
-	title = {Modeling of Polar Systems with the Perturbed-Chain SAFT Equation of State. Investigation of the Performance of Two Polar Terms},
-	volume = {44},
-	doi = {10.1021/ie050071c},
-	language = {en},
-	number = {17},
-	journal = {Industrial \& Engineering Chemistry Research},
-	author = {Dominik, Aleksandra and Chapman, Walter G. and Kleiner, Matthias and Sadowski, Gabriele},
-	year = {2005},
-	pages = {6928},
-}
-
 @article{tihic_applications_2006,
 	title = {Applications of the simplified perturbed-chain SAFT equation of state using an extended parameter table},
 	volume = {248},
