### 2025-02-06 #7365 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
